optimization
Перенос модулей в папку modules. Убрал ненужный модель datetime.
