+++ options.tardis
@@ -1,1 +1,1 @@
-{"name": "Пользователь", "assname": "ТАРДИС"}
+{"name": "artegoser", "assname": "\u0422\u0410\u0420\u0414\u0418\u0421"}
